Revert "Enable legacy MMAP"
We needed this for rmt_storage blob, but for serrano we now
uses the blob from deb, thus this is not needed.

Other msm8930 devices should be able to use the same blob,
but if they want to stay on the old one, they can set this
flag in their device config.

This reverts commit 1ceb258fe0e5311cdf56684809572c4bdeff3c2d.

Change-Id: I7e7633f527d7053af841813d4ff60fda359a925d